requirements.txt
# General dependencies
torch>=1.12.1
librosa>=0.9.2
numpy>=1.23.1
dataclasses
transformers>=4.20.1
datasets>=2.4.0
evaluate>=0.2.2
pyyaml
scipy>=1.9.1
tqdm>=4.64.0
jiwer
gradio
accelerate

# Specific versions and sources
bitsandbytes==0.41.1
git+https://github.com/huggingface/peft.git@main
